[dynamo] Log all compilation metrics in torch._dynamo.metrics
Motivation: for testing in dynamo, it would be nice to have an equivalent to torch._inductor.metrics. We already have log_compilation_event, but that only prints logs or dumps to a database (in fbcode); these aren't appropriate for unit tests.

This change: adds torch._dynamo.metrics, and then changes convert_frame so that it calls torch._dynamo.metrics.record_compilation_metrics instead. To prevent a super large list from forming, use a deque with a max length; then if the max length is exceeded, the oldest records will be dropped.

ghstack-source-id: ea76a5b41d6c367a7d16bea2cd147f925f21a1e6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115788